Cleaned up!  fixed tmp file.
